Update object-basics1-download.html
In the code update, I implemented bracket notation `cat['name']` to retrieve the value of the `name` property and assigned it to the `catName` variable. Additionally, I executed the `greeting()` method using dot notation `cat.greeting()`, resulting in the message 'Meow!' being logged to the console. 
Moreover, I utilized bracket notation `cat['color']` to modify the value of the `color` property to 'black'.